[kmac, fpga] Keep signaling idle until SHA3 starts in SCA mode
When doing SCA experiments, we want KMAC to keep signaling idle while
writing the message FIFO. The idle signal is used as capture trigger
it should only go low once the SHA3 core starts the actual operation.

This change is only enabled on the FPGA platform used for SCA
measurements.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>