[dv/shadow_reg] Shadow register write by field
This PR moves the committed_val, shaodw_val, and staged_val to
dv_base_reg_field because in design, they are updated by field.
We align in DV to avoid the corner case where shadow_storage_err field
with field 1, but writing valid value to field 2 still works in design.

Signed-off-by: Cindy Chen <chencindy@opentitan.org>