[ipgen] Strengthen template parameter type checks
The `IpConfig` class holds typed template parameters, which can be
either a string or an integer. Before, the parameter type was only
checked when data was loaded from text form (usually reading a Hjson
file), but not when constructing an `IpConfig` object programmatically.

Strengthen the checks to ensure that all template parameters which are
passed into `IpConfig` can ultimately be serialized to Hjson and survive
a round-trip.

This checking requires passing the list of expected template parameters,
encapsulated in `TemplateParams` around a bit more.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>