[chip, testplan] Indicate SW tests better
This is a simple search and replace of `chip_` in the test names with
`chip_sw_` in our chip level testplan as a way to indicated which tests
are SW based. Other alternative is to introduce additional tags / fields
into the testplan, which requires additional tooling enhancements. We
will address that later. For now, it should be ok to track SW tests with
`chip_sw` prefix.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>