[hw/aon_timer] Make IRQ and signal names consistent.
This fixes #8533.

Every IP that can raise IRQs has an `interrupt_list` field in its HJSON
file that lists the names (in snake case) and brief descriptions (in a
full sentence/phrase) of each interrupt it produces.

Additionally, for IPs that raise IRQs, their HJSON usually defines the
following three registers: INTR_STATE, INTR_ENABLE, and INTR_TEST.
Typically, the bit fields within each of those three registers map 1:1
with the interrupts listed above in the `interrupt_list` field, and
**_therefore their names match_**.

However, in the aon_timer, the `wdog_timer_bark` interrupt was also
referred to as `wdog_timer_expired`. This fixes this naming
irregularity.

Signed-off-by: Timothy Trippel <ttrippel@google.com>