[opentitantool] Add support for HyperDebug
Add a Transport implementation for the HyperDebug board.  The
implementation uses the rusb (Rust libusb) for finding the HyperDebug
USB device, and for bulk communication to do e.g. SPI communication.
Additionally, the impementation open the Linux /dev/ttyUSBn files
using a Rust serialport crate, which is used for serial communication
with the OPenTitan chip, as well as for sending generic text commands
to HyperDebug itself, to do e.g. GPIO-related operations.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: Id04c55d59da1f245da111c5a04f64f53cd2449a9