[otbn,dv] Zero mirrored STATUS register on reset
Without this change, the otbn_reset test fails. What happens is that a
running operation gets killed by a reset (leaving STATUS equal to 0x1)
and the mirrored STATUS value in the C++ code doesn't get zeroed,
which means a STATUS of 1 appears as soon as the model gets
re-started... one cycle too early.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>