[otbn,dv] Use clocking block to spot instruction execution
This fixes a problem caused by us switching to spotting STATUS changes
on the negedge of the clock (in commit bfe59f0a8).

The problem is triggered when executing a final ECALL instruction or
similar. The model immediately reports that STATUS gets cleared, which
is spotted on the next negedge of the clock. With some schedulings
from the simulator, we only see the ECALL instruction on the following
posedge, by which time the scoreboard thinks we've stopped execution.
This causes the check on line 290 of otbn_scoreboard.sv to explode.

Note that there's work afoot to flop the STATUS register. This problem
will go away anyway when that lands, but it seems cleaner to use a
clocking block to get rid of the race.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>