[otbn,dv] Directly signal an expected alert in IMEM error sequence
We were picking this up from the STATUS register, but commit 6386c75
has delayed that by a cycle, so we don't spot the impending error in
time. Arguably, that wasn't really the right way to do it anyway,
since we should be setting up an expected alert when we see the bad
thing happen, not a response to it.

Copy the clkmgr DV environment and poke the scoreboard directly from
the vseq. Another way to do things would be to monitor a "fatal error
interface", but that seems a bit round-about.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>