[rstmgr] Fix width of counter in rstmgr_cnsty_chk
The definition of CntWidth wasn't correct for all values of
MaxSyncDelay. In particular, it wouldn't work properly for a
MaxSyncDelay of 3 (where representing the next value takes another
bit).

Also, add an explicit cast to avoid a width mismatch warning. Now that
we've got the widths right, we could alternatively slice MaxSyncDelay.
But this seems slightly safer (and we've just got it wrong once!).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>